+++ src/playground/brain/brain.h
@@ -5,6 +5,6 @@
 
 extern inline void process() __attribute__((always_inline));
 extern inline struct pointer* get_full_path() __attribute__((always_inline));
-extern inline u32 get_file_size() __attribute__((always_inline));
+extern inline u64 get_file_size() __attribute__((always_inline));
 
 #endif // _PLAYGROUND_BRAIN_H_
